+++ src/runtime/main.ml
@@ -54,6 +54,18 @@ let force_start =
     ~comments:
       ["This should be reserved for advanced dynamic uses of liquidsoap."]
 
+(* Should we allow to run as root? *)
+let allow_root =
+  Dtools.Conf.bool
+    ~p:(Dtools.Init.conf#plug "allow_root")
+    ~d:(try Sys.file_exists "/.dockerenv" with _ -> false)
+    "Allow liquidsoap to run as root"
+    ~comments:
+      [
+        "This should be reserved for advanced dynamic uses of liquidsoap ";
+        "such as running inside an isolated environment like docker.";
+      ]
+
 (* Do not run, don't even check the scripts. *)
 let parse_only = ref false
 
@@ -500,6 +512,10 @@ let final_cleanup () =
   Gc.full_major ();
   Gc.full_major ()
 
+let sync_cleanup () =
+  initial_cleanup ();
+  final_cleanup ()
+
 let () =
   (* Shutdown *)
   Lifecycle.before_core_shutdown (fun () -> log#important "Shutdown started!");
@@ -579,10 +595,24 @@ let () =
                  Runtime.interactive ();
                  Tutils.shutdown 0)
                ());
-          main ())
+          Dtools.Init.init main)
         else if Source.has_outputs () || force_start#get then (
           check_directories ();
-          main ())
+          let msg_of_err = function
+            | `User -> "root euid (user)"
+            | `Group -> "root guid (group)"
+            | `Both -> "root euid & guid (user & group)"
+          in
+          let on_error e =
+            Printf.eprintf
+              "init: security exit, %s. Override with \
+               settings.init.allow_root.set(true)\n"
+              (msg_of_err e);
+            sync_cleanup ();
+            exit (-1)
+          in
+          try Dtools.Init.init ~prohibit_root:(not allow_root#get) main
+          with Dtools.Init.Root_prohibited e -> on_error e)
         else (
           final_cleanup ();
           Printf.printf "No output defined, nothing to do.\n";